`size_t` in `mojo::DataPipe[Consumer|Producer]Handle`: /google_apis.
This CL makes use of the new `size_t`-based APIs in the remaining
directories like `//components`, `//extensions`, or `//storage`.
See https://crrev.com/c/5414148 for motivation.

This CL may have introduced new integer conversions - in those cases the
CL attempts to have well-defined behavior (e.g. using
`base::checked_cast`).  OTOH, this CL did *not* attempt to fix old
integer operations (e.g. old `int` => `uint32_t` / `size_t` conversions,
or old additions that may potentially overflow).

This CL was uploaded by git cl split.

R=peter@chromium.org

Bug: 40284755
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5462641
Auto-Submit: Łukasz Anforowicz <lukasza@chromium.org>
Commit-Queue: Peter Beverloo <peter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1290600}

+++ google_apis/gcm/base/socket_stream.cc
@@ -11,6 +11,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 #include "base/functional/bind.h"
 #include "base/functional/callback.h"
+#include "base/numerics/safe_conversions.h"
 #include "net/base/io_buffer.h"
 #include "net/socket/stream_socket.h"
 
@@ -109,7 +110,7 @@ net::Error SocketInputStream::Refresh(base::OnceClosure callback,
     return net::OK;
   }
 
-  read_size_ = byte_limit;
+  read_size_ = base::checked_cast<size_t>(byte_limit);
   read_callback_ = std::move(callback);
   stream_watcher_.ArmOrNotify();
   last_error_ = net::ERR_IO_PENDING;
@@ -122,7 +123,7 @@ void SocketInputStream::ReadMore(
   DCHECK(read_callback_);
   DCHECK_NE(0u, read_size_);
 
-  uint32_t num_bytes = read_size_;
+  size_t num_bytes = read_size_;
   if (result == MOJO_RESULT_OK) {
     DVLOG(1) << "Refreshing input stream, limit of " << num_bytes << " bytes.";
     result = stream_->ReadData(read_buffer_->data(), &num_bytes,
@@ -155,7 +156,7 @@ void SocketInputStream::ReadMore(
     return;
 
   last_error_ = net::OK;
-  read_buffer_->DidConsume(num_bytes);
+  read_buffer_->DidConsume(base::checked_cast<uint32_t>(num_bytes));
   // TODO(zea): investigating crbug.com/409985
   CHECK_GT(UnreadByteCount(), 0);
 
@@ -289,7 +290,8 @@ void SocketOutputStream::WriteMore(MojoResult result,
   DCHECK(write_callback_);
   DCHECK(write_buffer_);
 
-  uint32_t num_bytes = write_buffer_->BytesRemaining();
+  size_t num_bytes =
+      base::checked_cast<size_t>(write_buffer_->BytesRemaining());
   DVLOG(1) << "Flushing " << num_bytes << " bytes into socket.";
   if (result == MOJO_RESULT_OK) {
     result = stream_->WriteData(write_buffer_->data(), &num_bytes,
@@ -313,7 +315,7 @@ void SocketOutputStream::WriteMore(MojoResult result,
 
   DCHECK_GE(num_bytes, 0u);
   last_error_ = net::OK;
-  write_buffer_->DidConsume(num_bytes);
+  write_buffer_->DidConsume(base::checked_cast<uint32_t>(num_bytes));
   if (write_buffer_->BytesRemaining() > 0) {
     DVLOG(1) << "Partial flush complete. Retrying.";
     // Only a partial write was completed. Flush again to finish the write.
